Minor updates BillingRepository documentation. Fixes #186
Without intending to be overly sensitive, there were some statements in
the documentation that didn't add value from a documentation standpoint
and can easily be viewed as condescending or unprofessional, especially
to frustrated users to are viewing the sample code to try to solve some
problems or shortcomings in the developer documentation.

This change updates references in the documentation to assumptions about
the reader's skill level, things they should or should not know, and
things that should or should not be difficult for them.