Do not hard code response code for RESULT_CANCELED
When an Activity results in an `Activity.RESULT_CANCELED` result code, it is often because the user pressed Back or Cancel. However, it is not the case 100% of the time. Instead of calling the listener with a hardcoded `IABHELPER_USER_CANCELLED` code, we should call it with the actual responseCode from the Intent.